src/utils/map-generator/pyproject.toml
[project]
name = "k"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mapbox-vector-tile>=2.2.0",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
